+++ userprofile/tests/test_views.py
@@ -67,6 +67,24 @@ class TestUserSettingsForm(AppSettingMixin, TestViewsBase):
             user=self.user,
         )
 
+        # Init test setting with options
+        self.setting_str_options = self._make_setting(
+            app_name=EXAMPLE_APP_NAME,
+            name='user_str_setting_options',
+            setting_type='STRING',
+            value='string1',
+            user=self.user,
+        )
+
+        # Init integer setting with options
+        self.setting_int_options = self._make_setting(
+            app_name=EXAMPLE_APP_NAME,
+            name='user_int_setting_options',
+            setting_type='INTEGER',
+            value=0,
+            user=self.user,
+        )
+
         # Init boolean setting
         self.setting_bool = self._make_setting(
             app_name=EXAMPLE_APP_NAME,
@@ -104,6 +122,16 @@ class TestUserSettingsForm(AppSettingMixin, TestViewsBase):
         )
         self.assertIsNotNone(
             response.context['form'].fields.get(
+                'settings.%s.user_str_setting_options' % EXAMPLE_APP_NAME
+            )
+        )
+        self.assertIsNotNone(
+            response.context['form'].fields.get(
+                'settings.%s.user_int_setting_options' % EXAMPLE_APP_NAME
+            )
+        )
+        self.assertIsNotNone(
+            response.context['form'].fields.get(
                 'settings.%s.user_bool_setting' % EXAMPLE_APP_NAME
             )
         )
@@ -129,6 +157,18 @@ class TestUserSettingsForm(AppSettingMixin, TestViewsBase):
         )
         self.assertEqual(
             app_settings.get_app_setting(
+                EXAMPLE_APP_NAME, 'user_str_setting_options', user=self.user
+            ),
+            'string1',
+        )
+        self.assertEqual(
+            app_settings.get_app_setting(
+                EXAMPLE_APP_NAME, 'user_int_setting_options', user=self.user
+            ),
+            0,
+        )
+        self.assertEqual(
+            app_settings.get_app_setting(
                 EXAMPLE_APP_NAME, 'user_bool_setting', user=self.user
             ),
             True,
@@ -143,6 +183,9 @@ class TestUserSettingsForm(AppSettingMixin, TestViewsBase):
         values = {
             'settings.%s.user_str_setting' % EXAMPLE_APP_NAME: 'another-text',
             'settings.%s.user_int_setting' % EXAMPLE_APP_NAME: '123',
+            'settings.%s.user_str_setting_options'
+            % EXAMPLE_APP_NAME: 'string2',
+            'settings.%s.user_int_setting_options' % EXAMPLE_APP_NAME: 1,
             'settings.%s.user_bool_setting' % EXAMPLE_APP_NAME: False,
             'settings.%s.user_json_setting'
             % EXAMPLE_APP_NAME: '{"Test": "Less"}',
@@ -169,6 +212,18 @@ class TestUserSettingsForm(AppSettingMixin, TestViewsBase):
                 EXAMPLE_APP_NAME, 'user_int_setting', user=self.user
             ),
             123,
+        )
+        self.assertEqual(
+            app_settings.get_app_setting(
+                EXAMPLE_APP_NAME, 'user_str_setting_options', user=self.user
+            ),
+            'string2',
+        )
+        self.assertEqual(
+            app_settings.get_app_setting(
+                EXAMPLE_APP_NAME, 'user_int_setting_options', user=self.user
+            ),
+            1,
         )
         self.assertEqual(
             app_settings.get_app_setting(